requirements.txt
streamlit
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
arxiv
wikipedia-api
duckduckgo-search
youtube-transcript-api
python-dotenv
watchdog